pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.15.0",
    "nest-asyncio>=1.6.0",
    "openai>=2.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
